+++ src/ecdsa/ot_based_ecdsa/mod.rs
@@ -1,3 +1,41 @@
+
+use serde::{Deserialize, Serialize};
+use triples::{TriplePub, TripleShare};
+use crate::ecdsa::{
+    KeygenOutput,
+    AffinePoint,
+    Scalar,
+};
+
+/// The output of the presigning protocol.
+///
+/// This output is basically all the parts of the signature that we can perform
+/// without knowing the message.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct PresignOutput {
+    /// The public nonce commitment.
+    pub big_r: AffinePoint,
+    /// Our share of the nonce value.
+    pub k: Scalar,
+    /// Our share of the sigma value.
+    pub sigma: Scalar,
+}
+
+/// The arguments needed to create a presignature.
+#[derive(Debug, Clone)]
+pub struct PresignArguments {
+    /// The first triple's public information, and our share.
+    pub triple0: (TripleShare, TriplePub),
+    /// Ditto, for the second triple.
+    pub triple1: (TripleShare, TriplePub),
+    /// The output of key generation, i.e. our share of the secret key, and the public key package.
+    /// This is of type KeygenOutput<Secp256K1Sha256> from Frost implementation
+    pub keygen_out: KeygenOutput,
+    /// The desired threshold for the presignature, which must match the original threshold
+    pub threshold: usize,
+}
+
+
 pub mod triples;
 pub mod presign;
 pub mod sign;
